# Use non-contiguous buffer for codec and transport